将安装过程存入log

+++ initialize.sh
@@ -8,6 +8,12 @@ fi
 # 切换到用户目录
 cd ~
 touch initialize-script.log
+log_file=initialize-script.log
+if [ -f $log_file ];then
+ echo "Warning: You have already run this script"
+ exit 1
+fi
+
 # 安装第三方源
 echo "Installing third-party sources..."
 yum -y install epel-release >> initialize-script.log
